Updates after review, Use module fixture to mock external requests
